Remove AspectMock usage from PostContentTransformerTest [MAILPOET-2454]
